requirements.txt
# requirements.txt
fastmcp>=0.1.0
manim>=0.18.0